onboarding_simple_tech/requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
colorgram.py==1.2.0
colorthief==0.2.1
comm==0.1.3
contourpy==1.0.7
cryptography==42.0.2
cycler==0.11.0
databases==0.8.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==5.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
dnspython==2.3.0
dominate==2.7.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.109.0
fastjsonschema==2.17.1
filelock==3.12.4
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.0
greenlet==2.0.2
gTTS==2.4.0
h11==0.14.0
idna==3.4
ipykernel==6.23.3
ipython==8.14.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
LinearRegression==0.0.1
looseversion==1.3.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
multidict==6.0.4
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.23.5
opencv-python==4.8.1.78
outcome==1.2.0
overrides==7.3.1
packaging==23.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
patsy==0.5.3
pdfminer.six==20221105
pdfplumber==0.10.3
pickleshare==0.7.5
Pillow==9.4.0
pipenv==2023.9.8
platformdirs==3.10.0
plotly==5.15.0
plotly-express==0.4.1
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2==2.9.9
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pygame==2.5.2
Pygments==2.15.1
PyJWT==2.8.0
PyMuPDF==1.23.21
PyMuPDFb==1.23.9
pyparsing==3.0.9
PyPDF2==3.0.1
pypdfium2==4.26.0
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2022.6
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
redis==4.5.3
regex==2023.8.8
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.8.3
Send2Trash==1.8.2
six==1.16.0
sklearn==0.0.post5
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spotipy==2.22.1
SQLAlchemy==1.4.51
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.35.1
statsmodels==0.14.0
tenacity==8.2.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
twilio==8.9.1
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
urllib3==1.26.14
uvicorn==0.27.0.post1
virtualenv==20.24.5
visitor==0.1.3
waitress==3.0.0
wcwidth==0.2.6
webcolors==1.13
webdriver-manager==3.8.5
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
yarl==1.9.2
